requirements.txt
altair
google-cloud-bigquery 
google-cloud-bigquery-storage
opentelemetry-api 
opentelemetry-instrumentation
pandas
tqdm
scikit-learn
scipy
